requirements/dev.txt
-r base.txt
ipython==8.25.0
ipdb==0.13.13
psycopg==3.1.19
psycopg-binary==3.1.19
djangorestframework==3.15.2
django-extensions==3.2.3